+++ extensions/gif-search/CHANGELOG.md
@@ -1,9 +1,10 @@
 # GIF Search Changelog
 
-## [Add unified GIF search command & adding to favorites] - 2022-04-07
+## [Add unified GIF search command & adding to favorites] - 2022-04-09
 - Adds a unified "Search for GIFs" command that searches across GIF services
 - Use gif slug as file name when copying file to clipboard
 - Add action for adding a gif to your favorites
+- Mark favorite GIFs with Star Icon in results list
 
 ## [Remove static GIF preview warning 🎉] - 2022-04-06
 - Removes the warning about GIFs no longer animating, Raycast v1.32.0 supports GIFs in Details